Start of Cron, fixed jobstore, add pause and resume, split task,
